requirements.txt
# AthenaGuard - Geminiathon Edition
# Unified security defense system with Gemini AI

# === Core Dependencies ===
openai>=1.0.0          # Gemini API via OpenAI-compatible endpoint
anthropic>=0.39.0      # Claude API for judge (optional)
pydantic>=2.0.0        # Schemas and validation
python-dotenv>=1.0.0   # Environment management

# === Web/API ===
fastapi>=0.115.0       # Judge server
uvicorn>=0.32.0        # ASGI server
httpx>=0.27.0          # Async HTTP client

# === MCP (Model Context Protocol) ===
mcp>=1.1.2             # MCP server/client

# === Security Scanning ===
semgrep>=1.0.0         # SAST scanning

# === Testing ===
pytest>=8.0.0          # Test runner
pytest-asyncio>=0.23.0 # Async test support

# === CLI & UI ===
rich>=13.0.0           # Rich terminal output
typer>=0.12.0          # CLI framework

# === Utilities ===
aiohttp>=3.9.0         # Async HTTP
tenacity>=8.2.0        # Retry logic

# === Training (Optional - for RL) ===
# Uncomment if running full training:
# openpipe-art>=0.5.0
# wandb>=0.16.0
# litellm>=1.30.0
# huggingface_hub>=0.20.0
# transformers>=4.40.0
# torch>=2.0.0

# === Development ===
ruff>=0.1.0            # Linter (optional)
